requirements.txt
Flask
Flask_SQLAlchemy
Flask_Migrate
Flask_Bootstrap
python-dotenv
PyMySQL
pandas
openpyxl
cryptography
pdfkit
